updateArr now only takes in the array memory and returns the array memory instead of returning the whole memory.
Added a lexer, parser and AST file for the input so the user can initialize variables through console.

Added 4 methods: initializeVar, initializeArr, setupArrAsList and initializeMemory to load/setup the
initial memory based on the parsed input for initialization of variables and arrays.

Problem when running the Calculator.fsx file the following error is given:
error FS0039: The namespace or module 'InputParser' is not defined. Maybe you want one of the following: InputTypesAST

I can't figure out why.